Fix broken UI issue
For the second ad pod, the UI was freezing
after the skip card. This was due to a bug
in the video player. When we don't skip at
least ~2 seconds past the end of the ad
break, the user will be stuck with a frozen
UI. With a two second additional skip-ahead
this issue does not reproduce.